Added a (rudimentary) python install script.
This script will go through all of the files in the repo (ignoring
anything specified in .gitignore, as well as a few others (such as the
README)), and create symlinks to them in the user's home directory.

Use with CAUTION - every time this script is run, it overwrites any
clashing files already in the home directory.