Added option to back up existing files.
If the Python install script detects that a file it is trying to link
already exists in the home directory, it will ask if you want to backup
that file (by adding successive tildes to the end of the filename,
until it is unique).

Also cleaned up the imports in the file and added some helpful
section-dividing comments.