Fixed Enthought Python path setting.
Previously, the option in .path to add Enthough Python to $PATH added
it straight to the end of the variable - meaning that it automatically
got overridden by the homebrew python.

This is now fixed, the Enthough Python bin gets added to the $PATH
right at the start, and so overrides any other python executables that
might be found further down the PATH variable.