Fixed default behaviour of j and k keys.
By default the `j` and `k` keys perform their actions upon the lines in the file (i.e, those separated by literal newline characters). This is often not what is desired, as wrapped lines are much harder to navigate around.

The remapping of `j` and `k` to `gj` and `gk` (respectively) allows them to navigate the lines on the screen, not just the lines in the file.