Added functionality to pass 0, 1 or more files to the commands.
The files are specified by name only, and the system then calculates the
full paths to each file (using the directory that the script is in as
the assumed dotfiles repository).

If 0 files are passed, all valid dotfiles in the repo are used as a
default value. This also ignores all files specified in the .gitignore
and in the IGNORED_FILES list at the top of the script.