`list` now checks to see where a link points
Before, the `list` command would just check to see if there was a symbolic link in the user's home directory with the same name as a file in the dot files repo. Whilst this will usually work, it can't be relied upon.

The `list` command will now check to see if there is an appropriate link in the home directory, as before, but will now also check to ensure the link points at the corresponding file in the dot files repo.