zsh: Move .zshenv to .zprofile
I noticed that `brew doctor` was complaining that /usr/local/bin was
earlier in the $PATH than /opt/homebrew/bin - despite my .zshenv file
prepending that to the $PATH.

It turns out that the source order for ZSH startup is:
  1. /etc/zshenv
  2. ~/.zshenv
  3. /etc/zprofile
  4. ~/.zprofile
  5. /etc/zshrc
  6. ~/.zshrc
  7. /etc/zlogin
  8. ~/.zlogin

I was prepending to my $PATH variable in 2. whilst 3., the default Mac
OS /etc/zprofile, was further prepending to it afterwards.