nvim: Add explore files option to starter dashboard
I'm starting to think that this dashboard isn't really what I want.
Instead of having to remember the letters that different actions are
mapped to in the dashboard, I want to be able to just use the regular
vim mappings that are already in my memory.

Basically the only things I want from a dashboard are:
 - A minimal, nice-to-look-at starting screen, with `future` output
 - A list of recently edited files and a quick way to jump into them

For now I've added "- Explore files" as an option to the dashboard, so
that I can at least still use the same `-` key as everywhere else to
open the file explorer.