nvim: Configure bufferline plugin
I'm not sure I'm totally happy with this. I'm pulling it in for now as
it's what I've been using in LazyVim, and a couple of the keymaps are
handy - but longer-term I might start swapping this out for some
combination of the telescope buffer picker (currently at <leader>,) or
Harpoon.