tmux: Swap prefix to ctrl+s instead of ctrl+f
I'm starting to wish for ctrl+f in vim more and more - it's used for
scrolling buffers, scrolling documentation floats, scrolling in
telescope, etc. Double-tapping ctrl+f to move forward one screen just
isn't cutting it.

The classic tmux binding of ctrl+b is a bit too awkward to type, and I
find the most popular alternative of ctrl+a too annoying, as I use
ctrl+a to jump to the start of a line in shells all the time.

It's actually quite hard to find a ctrl key combination that doesn't
have a conflicting mapping in vim or zsh. ctrl+s is used to freeze
screen output in zsh (which can then be resumed again with ctrl+q) -
which isn't a feature I ever use on purpose, so it seems ok to overwrite
it here.